modif

+++ PeerDiscovery.py
@@ -14,7 +14,7 @@ class PeerDiscovery:
 
     def load_peers(self):
         if os.path.isfile(seeds_path):
-            with open(seeds_path, "wb+") as f:
+            with open(seeds_path, "w+") as f:
                 self.peers = json.load(f)
 
     def peer_ip(self):
@@ -22,8 +22,12 @@ class PeerDiscovery:
 
     def add_peer(self, data):
         self.peers.append(data)
-        with open(seeds_path, "wb+") as f:
-            json.dump(self.peers, f)
+        if os.path.exists(seeds_path):
+            with open(seeds_path, "r") as f:
+                json.dump(self.peers, f)
+        else:
+            with open(seeds_path, "w+") as f:
+                json.dump(self.peers, f)
 
     def peer_register(self, message):
         payload = message.payload
